ecommerce-api/requirements.txt
flask
psycopg2-binary 
redis 
python-dotenv 
flask-sqlalchemy
